feat: quantification

+++ src/read_param.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 /**
  * @brief Lis les paramètre d'entré du programme et modifie des variables global 
@@ -9,7 +10,7 @@
  * @param argc 
  * @param argv 
  */
-void read_param(int argc, char *argv[], int *H1, int *V1, int *H2, int *V2, int *H3, int *V3, char **OUTFILE, char **INFILE){
+void read_param(int argc, char *argv[], uint8_t *H1, uint8_t *V1, uint8_t *H2, uint8_t *V2, uint8_t *H3, uint8_t *V3, char **OUTFILE, char **INFILE){
 
     int i = 1;
